requirements.txt
numpy>=2.0.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
matplotlib>=3.8.0
plotly>=5.22.0
numba>=0.60.0
llvmlite>=0.43.0
